add new class DataFinder
